+++ src/main/java/com/avixy/qrtoken/negocio/servico/Service.java
@@ -1,15 +1,18 @@
 package com.avixy.qrtoken.negocio.servico;
 
-import java.util.Map;
-
 /**
- * Created by I7 on 31/07/2014.
+ * Created on 31/07/2014
+ * @author Breno Salgado <breno.salgado@avixy.com>
  */
 public interface Service {
     public String getServiceName();
 
-    public byte getServiceCode();
+    public int getServiceCode();
 
+    /**
+     *
+     * @return Tripa de dados que representam os parâmetros do serviço conforme definido no protocolo de serviços
+     */
     byte[] getData();
 
 }
